Invoke certificate callback upon https connection
The change in https://android-review.googlesource.com/c/platform/packages/services/BuiltInPrintService/+/831249
requires notification by libcups to any registered server_cert_cb callback.
This addition obtains the X509 certificate public key, wraps it in a CUPS
array, and delivers it. If the callback exists and returns an error (meaning
the certificate was rejected) the communication fails.

Test: See related change above.

Change-Id: I64f0794af392eb883ba48c46a9bf79ccbca72f66
Signed-off-by: Glade Diviney <mopriadevteam@gmail.com>